requirements.txt
discord.py
aiohttp
pdfminer.six
pytesseract
Pillow
pdf2image
